Change u64 to usize in un/register_event
We can safely use usize instead of u64, because normally we will
not register a data that exceeds the size of usize

+++ vhost-user-backend/src/handler.rs
@@ -222,7 +222,7 @@ where
                         if let Err(e) = self.handlers[thread_index].register_event(
                             fd.as_raw_fd(),
                             EventSet::IN,
-                            u64::from(evt_idx),
+                            evt_idx as usize,
                         ) {
                             if e.kind() != io::ErrorKind::AlreadyExists {
                                 // This could happen if we're asked by the frontend to enable an
@@ -234,7 +234,7 @@ where
                         let _ = self.handlers[thread_index].unregister_event(
                             fd.as_raw_fd(),
                             EventSet::IN,
-                            u64::from(evt_idx),
+                            evt_idx as usize,
                         );
                     }
                     break;
